requirements.txt
# Core RAG stack
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-chroma==0.1.4
langchain-ollama==0.1.3
langchain-text-splitters==0.2.4

# Vector store
chromadb==0.5.3

# HTTP requests (for web ingestion)
requests==2.32.3

# Config
pyyaml==6.0.2

# Optional: OpenAI support
# langchain-openai==0.1.23
# openai==1.40.0

# Testing
pytest==8.3.2
pytest-cov==5.0.0

# Notebook support (optional)
jupyter==1.1.1
ipykernel==6.29.5
